requirements.txt
fastapi
uvicorn
openai-whisper
torch
termcolor
websockets 